fix update and delete animals

+++ animal-microservice/controllers/AnimalController.js
@@ -2,6 +2,15 @@ const animalService = require("../services/AnimalService");
 const axios = require("axios");
 const gatewayPort = process.env.GATEWAY_PORT || 3000;
 
+
+function get_user_info_from_token(token) {
+  if (token.startsWith("Bearer ")) {
+    token = token.slice(7, token.length);
+  }
+  const decoded = jwt.verify(token, JWT_SECRET);
+  return { userId: decoded.user_id, roleId: decoded.role_id };
+}
+
 exports.getAllAnimals = async (req, res) => {
   try {
     const animals = await animalService.getAllAnimals();
@@ -62,9 +71,25 @@ exports.getAnimalById = async (req, res) => {
 
 exports.updateAnimal = async (req, res) => {
   try {
-    await animalService.updateAnimal(req.params.id, req.body);
-    const animal = await animalService.getAnimalById(req.params.id);
-    res.json({ data: animal, status: "Success" });
+    const role_info = await axios.get(
+      `http://localhost:${gatewayPort}/auth/getRole`,{
+        headers: {
+          authorization: req.headers.authorization
+        }
+      }
+    );
+    if (role_info.data.role === "admin"){
+      const animal = await animalService.getAnimalById(req.params.id);
+      await animalService.updateAnimal(req.params.id, req.body);
+      return res.json({ data: animal, status: "Success" });
+    } else if (role_info.data.role === "user"){
+      user_id = get_user_info_from_token(req.headers.authorization).userId;
+      const animal = await animalService.getAnimalById(req.params.id);
+      if (user_id === animal.user_id){
+        await animalService.updateAnimal(req.params.id, req.body);
+        return res.json({ data: animal, status: "Success" });
+      }
+    }
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
@@ -72,9 +97,25 @@ exports.updateAnimal = async (req, res) => {
 
 exports.deleteAnimal = async (req, res) => {
   try {
-    const animal = await animalService.getAnimalById(req.params.id);
-    await animalService.deleteAnimal(req.params.id);
-    res.json({ data: animal, status: "Successfully deleted" });
+    const role_info = await axios.get(
+      `http://localhost:${gatewayPort}/auth/getRole`,{
+        headers: {
+          authorization: req.headers.authorization
+        }
+      }
+    );
+    if (role_info.data.role === "admin"){
+      const animal = await animalService.getAnimalById(req.params.id);
+      await animalService.deleteAnimal(req.params.id);
+      return res.json({ data: animal, status: "Successfully deleted" });
+    }else if (role_info.data.role === "user") {
+      user_id = get_user_info_from_token(req.headers.authorization).userId;
+      const animal = await animalService.getAnimalById(req.params.id);
+      if (user_id === animal.user_id){
+        await animalService.deleteAnimal(req.params.id);
+        return res.json({ data: animal, status: "Successfully deleted" });
+      }
+    }
   } catch (err) {
     res.status(500).json({ error: err.message });
   }
